Add seach subcommand to display OTP codes for matching db entries
This commit adds a `search` subcommand that will search the database
for elements matching either the issuer field, the label field, or
both. An OTP code for each matching element will be generated and
the issuer, label and generated code with time remaining
will be printed.

The `search` subcommand takes -i, --issuer and -l, --label flags.
At least one of the two is required. If both are provided then only
elements with fields matching both arguments will be printed.